requirements.txt
#libraries
streamlit
google-generativeai
pyhton-dotenv